ecofinds2/ecofinds/ecofinds_project/requirements.txt
# EcoFinds - Production Requirements
# Core Django Framework
Django==5.2.6
djangorestframework==3.16.1
django-cors-headers==4.3.1

# Database
psycopg2-binary==2.9.9  # PostgreSQL adapter
dj-database-url==2.1.0

# Image Processing
Pillow==11.3.0
django-imagekit==4.1.0

# AI & Machine Learning
tensorflow==2.15.0
opencv-python==4.8.1.78
numpy==1.24.3
scikit-learn==1.3.2

# Translation & Localization
googletrans==4.0.0rc1
django-modeltranslation==0.18.11

# Real-time Features
channels==4.0.0
channels-redis==4.1.0
redis==5.0.1

# API Integration
requests==2.31.0
httpx==0.25.2

# Security
django-allauth==0.57.0
django-ratelimit==4.1.0
cryptography==41.0.7

# Monitoring & Logging
sentry-sdk==1.38.0
django-debug-toolbar==4.2.0

# File Storage
boto3==1.34.0  # AWS S3
django-storages==1.14.2

# Environment & Configuration
python-decouple==3.8
django-environ==0.11.2

# Testing
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0

# Development Tools
black==23.11.0
flake8==6.1.0
isort==5.12.0

# Production Server
gunicorn==21.2.0
whitenoise==6.6.0

# Additional Utilities
celery==5.3.4
django-celery-beat==2.5.0
python-dateutil==2.8.2
pytz==2023.3

